PagerView用の実習も共存できるようにMainActivityを置く
