+++ AndroidStudio/practice/fundamentals/6th/ListViewPractice/src/main/java/jp/mixi/practice/listview/MainActivity.java
@@ -1,31 +1,23 @@
-
 package jp.mixi.practice.listview;
 
+import android.content.Context;
+import android.content.Intent;
 import android.os.Bundle;
+import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
-import android.widget.ArrayAdapter;
-import android.widget.ListView;
-
-import java.util.ArrayList;
-import java.util.List;
+import android.view.View;
 
 public class MainActivity extends AppCompatActivity {
-
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-        // データの作成
-        List<String> listData = new ArrayList<>();
-        for (int i = 0; i < 20; i++) {
-            listData.add("タイトル" + i);
-        }
-
-        ArrayAdapter<String> adapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_1, listData);
-
-        ListView listView = (ListView) findViewById(R.id.list_view);
-        //TODO ここでセットするadapterがCustomListItemAdapterになるように変更してください
-        listView.setAdapter(adapter);
+        findViewById(R.id.list_view1_button).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Context context = view.getContext();
+                context.startActivity(new Intent(context, ListViewActivity.class));
+            }
+        });
     }
 }
